Major advancements
Made a shell script to facilitate compiling.

Moved most code from the main class to other files and changed the main while loop to use the handle() method which will be able to parse errors.

Made a basic command parser -- and it works!

Made a small file to map error codes.

Built a basic command input system.

This is big advancements: fun!