Functional command system
Somewhat rewrote the `.comple.sh` for a brighter future.

Moved most of the `input_handler()` function to other functions and other files for easier readability.

Separated the screen into multiple windows, this will hopefully be the next commit.

Overall, this commit makes the command system truly working with a command line, output, and a functional echo command to view environment variables and, later on, program variables.